requirements.txt
asgiref==3.7.2
Django==4.2.6
passlib==1.7.4
psycopg2-binary==2.9.9
sqlparse==0.4.4
typing-extensions==4.8.0
djangorestframework
python-dotenv